+++ src/components/FormatSelector.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Select } from '@grafana/ui';
-import { InfinityQuery, SCRAP_QUERY_RESULT_FORMATS, InfinityQueryFormat } from '../types';
+import { InfinityQuery, INFINITY_RESULT_FORMATS, InfinityQueryFormat } from '../types';
 import { Components } from '../selectors';
 import { isDataQuery } from 'app/utils';
 interface FormatSelectorProps {
@@ -10,7 +10,7 @@ interface FormatSelectorProps {
 }
 export const FormatSelector = (props: FormatSelectorProps) => {
   const { query, onChange, onRunQuery } = props;
-  if (!isDataQuery(query)) {
+  if (!(isDataQuery(query) || query.type === 'uql')) {
     return <></>;
   }
   const onFormatChange = (format: InfinityQueryFormat) => {
@@ -19,9 +19,11 @@ export const FormatSelector = (props: FormatSelectorProps) => {
   };
   const getFormats = () => {
     if (query.type === 'json' && query.source === 'inline') {
-      return SCRAP_QUERY_RESULT_FORMATS;
+      return INFINITY_RESULT_FORMATS;
+    } else if (query.type === 'uql') {
+      return INFINITY_RESULT_FORMATS.filter((f) => f.value === 'table' || f.value === 'timeseries' || f.value === 'dataframe');
     } else {
-      return SCRAP_QUERY_RESULT_FORMATS.filter((f) => f.value !== 'as-is');
+      return INFINITY_RESULT_FORMATS.filter((f) => f.value !== 'as-is');
     }
   };
   return (