+++ src/editors/query/query.url.tsx
@@ -1,22 +1,18 @@
 import React, { useState } from 'react';
 import { InlineFormLabel, Icon } from '@grafana/ui';
 import { URLOptionsEditor } from './query.url.options';
+import { isDataQuery } from './../../app/utils';
 import { InfinityQuery, EditorMode } from '../../types';
-import { isDataQuery } from 'app/utils';
 
-interface ScrapperProps {
-  query: InfinityQuery;
-  mode: EditorMode;
-  onChange: (value: any) => void;
-  onRunQuery: () => void;
-}
-
-export const URLEditor = (props: ScrapperProps) => {
+export const URLEditor = (props: { query: InfinityQuery; mode: EditorMode; onChange: (value: any) => void; onRunQuery: () => void }) => {
   const { query, onChange, onRunQuery } = props;
-  const canShowURLField = isDataQuery(query) && query.source === 'url';
   const LABEL_WIDTH = props.mode === 'variable' ? 10 : 8;
-  const [data, setData] = useState(isDataQuery(query) && query.source === 'inline' ? query.data || '' : '');
-  const [url, setURL] = useState(isDataQuery(query) && query.source === 'url' ? query.url || '' : '');
+  const canShowURLField = (isDataQuery(query) || query.type === 'uql') && query.source === 'url';
+  const [data, setData] = useState((isDataQuery(query) || query.type === 'uql') && query.source === 'inline' ? query.data || '' : '');
+  const [url, setURL] = useState((isDataQuery(query) || query.type === 'uql') && query.source === 'url' ? query.url || '' : '');
+  if (!(isDataQuery(query) || query.type === 'uql')) {
+    return <></>;
+  }
   const onDataChange = () => {
     onChange({ ...query, data, url });
     onRunQuery();
@@ -51,14 +47,7 @@ export const URLEditor = (props: ScrapperProps) => {
             Data
           </InlineFormLabel>
           <textarea rows={5} className="gf-form-input" style={{ width: '594px' }} value={data} placeholder="" onBlur={onDataChange} onChange={(e) => setData(e.target.value)}></textarea>
-          <Icon
-            name="play"
-            size="lg"
-            style={{ color: 'greenyellow' }}
-            onClick={() => {
-              onDataChange();
-            }}
-          />
+          <Icon name="play" size="lg" style={{ color: 'greenyellow' }} onClick={() => onDataChange()} />
         </div>
       )}
     </>